+++ changelog.md
@@ -5,6 +5,7 @@
 This version is in progress.
 
 * **Breaking**: built-in parsers explicitly leave string stripping in the control of the main `STRIP` constant. They don't strip additionally themselves.
+* Hash and Array parsers are built-in but not enabled by default.
 
 ## v0.1.1